@@ -20,7 +20,7 @@ namespace Mtd.OrderMaker.Server.Areas.Identity.Pages.Account.Manage
         private readonly UserManager<WebAppUser> _userManager;
         private readonly SignInManager<WebAppUser> _signInManager;
         private readonly IEmailSender _emailSender;
-        private readonly IHostingEnvironment _hostingEnvironment;
+        private readonly IWebHostEnvironment _hostingEnvironment;
         private readonly IOptions<ConfigSettings> _options;
         private readonly IStringLocalizer<IndexModel> _localizer;
 
@@ -29,7 +29,7 @@ namespace Mtd.OrderMaker.Server.Areas.Identity.Pages.Account.Manage
             UserManager<WebAppUser> userManager,
             SignInManager<WebAppUser> signInManager,
             IEmailSender emailSender, 
-            IHostingEnvironment hostingEnvironment,
+            IWebHostEnvironment hostingEnvironment,
             IStringLocalizer<IndexModel> localizer,
             IOptions<ConfigSettings> options
             )
